requirements.txt
asttokens==3.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fonttools==4.62.1
future==1.0.0
ifaddr==0.2.0
ipykernel==7.2.0
ipympl==0.10.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
iso8601==2.1.0
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
keyboard==0.13.5
kiwisolver==1.5.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.20.0
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
PyVISA==1.16.2
PyVISA-py==0.8.1
PyYAML==6.0.3
pyzmq==27.1.0
six==1.17.0
sockets==1.0.0
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.6.0
widgetsnbextension==4.0.15
zeroconf==0.148.0
